move group search to it's own folder
